made modules and prepared the files to work with a db but im a getting an error
